added coordiante geometry type

+++ src/graphql/modules/Location/location.type.graphql.ts
@@ -31,6 +31,6 @@ export default gql`
 
   type Geometry {
     type: String!
-    coordinates: [Int]!
+    coordinates: [Float]!
   }
 `
